# Add a "useProcFiles" option and speed up performance by keeping fd open