--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,7 @@
 sudo: false
 language: node_js
 node_js:
-  - '9'
+  - '10'
   - '8'
   - '6'
   - '4'