--- a/lib/helpers/parallel.js
+++ b/lib/helpers/parallel.js
@@ -9,13 +9,18 @@ function parallel (fns, options, done) {
 
   var keys
   if (!Array.isArray(fns)) { keys = Object.keys(fns) }
-  var pending = keys ? keys.length : fns.length
+  var length = keys ? keys.length : fns.length
+  var pending = length
   var results = keys ? {} : []
 
   function each (i, err, result) {
     results[i] = result
 
     if (--pending === 0 || (err && !options.graceful)) {
+      if (options.graceful && err && length > 1) {
+        err = null
+      }
+
       done && done(err, results)
       done = null
     }