--- a/lib/history.js
+++ b/lib/history.js
@@ -3,6 +3,7 @@ var DEFAULT_MAXAGE = 60000
 
 var expiration = {}
 var history = {}
+var expireListeners = {}
 
 var size = 0
 var interval = null
@@ -19,7 +20,7 @@ function get (pid, maxage) {
   return history[pid]
 }
 
-function set (pid, object, maxage) {
+function set (pid, object, maxage, onExpire) {
   if (object === undefined || maxage <= 0) return
 
   expiration[pid] = Date.now() + (maxage || DEFAULT_MAXAGE)
@@ -27,7 +28,11 @@ function set (pid, object, maxage) {
     size++
     sheduleInvalidator(maxage)
   }
+
   history[pid] = object
+  if (onExpire) {
+    expireListeners[pid] = onExpire
+  }
 }
 
 function sheduleInvalidator (maxage) {
@@ -49,10 +54,16 @@ function runInvalidator () {
   var now = Date.now()
   var pids = Object.keys(expiration)
   for (var i = 0; i < pids.length; i++) {
-    if (expiration[pids[i]] < now) {
+    var pid = pids[i]
+    if (expiration[pid] < now) {
       size--
-      delete history[pids[i]]
-      delete expiration[pids[i]]
+      if (expireListeners[pid]) {
+        expireListeners[pid](history[pid])
+      }
+
+      delete history[pid]
+      delete expiration[pid]
+      delete expireListeners[pid]
     }
   }
   sheduleInvalidator()