--- a/lib/ps.js
+++ b/lib/ps.js
@@ -2,6 +2,7 @@
 
 var os = require('os')
 var bin = require('./bin')
+var history = require('./history')
 
 var PLATFORM = os.platform()
 
@@ -35,7 +36,6 @@ function parseTime (timestr, fraction) {
   */
 function ps (pids, options, done) {
   var pArg = pids.join(',')
-
   var args = ['-o', 'etime,pid,ppid,pcpu,rss,time', '-p', pArg]
 
   if (PLATFORM === 'aix') {
@@ -93,11 +93,18 @@ function ps (pids, options, done) {
       }
 
       var pid = parseInt(line[1], 10)
+      var hst = history.get(pid, options.maxage)
+      if (hst === undefined) hst = {}
+
       var ppid = parseInt(line[2], 10)
-      var cpu = parseFloat(line[3].replace(',', '.'), 10)
       var memory = parseInt(line[4], 10) * 1024
       var etime = parseTime(line[0])
       var ctime = parseTime(line[5], true)
+
+      var total = (ctime - (hst.ctime || 0))
+      // time elapsed between calls in seconds
+      var seconds = Math.abs(hst.etime !== undefined ? etime - hst.etime : etime)
+      var cpu = seconds > 0 ? (total / seconds) * 100 : 0
 
       statistics[pid] = {
         cpu: cpu,
@@ -108,6 +115,8 @@ function ps (pids, options, done) {
         elapsed: etime,
         timestamp: date
       }
+
+      history.set(pid, statistics[pid], options.maxage)
     }
 
     done(null, statistics)