--- a/lib/stats.js
+++ b/lib/stats.js
@@ -5,23 +5,29 @@ var os = require('os')
 
 var platformToMethod = {
   darwin: 'ps',
-  sunos: 'ps',
-  freebsd: 'ps',
-  netbsd: 'ps',
+  sunos: 'procfile',
+  freebsd: 'procfile',
+  netbsd: 'procfile',
   win: 'wmic',
-  linux: 'ps',
-  aix: 'ps'
+  linux: 'procfile',
+  aix: 'procfile',
+  alpine: 'procfile'
 }
 
-var useProcfile = fs.existsSync('/etc/alpine-release')
+var ps = require('./ps')
 var platform = os.platform()
+
+if (fs.existsSync('/etc/alpine-release')) {
+  platform = 'alpine'
+}
+
 if (platform.match(/^win/)) {
   platform = 'win'
 }
 
 var stat
 try {
-  stat = useProcfile ? require('./procfile') : require('./' + platformToMethod[platform])
+  stat = require('./' + platformToMethod[platform])
 } catch (err) {}
 
 /**
@@ -38,6 +44,11 @@ try {
  * @param  {pidCallback} callback Called when the statistics are ready.
  */
 function get (pids, options, callback) {
+  var fn = stat
+  if (platform !== 'win' && options.usePs === true) {
+    fn = ps
+  }
+
   if (stat === undefined) {
     return callback(new Error(os.platform() + ' is not supported yet, please open an issue (https://github.com/soyuka/pidusage)'))
   }
@@ -59,7 +70,7 @@ function get (pids, options, callback) {
     }
   }
 
-  stat(pids, options, function (err, stats) {
+  fn(pids, options, function (err, stats) {
     if (err) {
       return callback(err)
     }