--- a/test/bench.js
+++ b/test/bench.js
@@ -33,13 +33,13 @@ async function destroy (childs) {
   childs.forEach(child => child.kill())
 }
 
-async function execute (childs, pidno, times) {
+async function execute (childs, pidno, times, options = {}) {
   var pids = childs.map(child => child.pid).slice(0, pidno)
 
   const end = tspan()
   try {
     for (let i = 0; i < times; i++) {
-      await m(pids)
+      await m(pids, options)
     }
     const time = end()
     return Promise.resolve(time)
@@ -55,23 +55,44 @@ test.serial('should execute the benchmark', async t => {
   let time = await execute(childs, 1, 100)
   t.log(`1 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
+  time = await execute(childs, 1, 100, {useProcFiles: true})
+  t.log(`(procfile) 1 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
   time = await execute(childs, 2, 100)
   t.log(`2 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
+  time = await execute(childs, 2, 100, {useProcFiles: true})
+  t.log(`(procfile) 2 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
   time = await execute(childs, 5, 100)
   t.log(`5 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
+  time = await execute(childs, 5, 100, {useProcFiles: true})
+  t.log(`(procfile) 5 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
   time = await execute(childs, 10, 100)
   t.log(`10 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
+  time = await execute(childs, 10, 100, {useProcFiles: true})
+  t.log(`(procfile) 10 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
   time = await execute(childs, 25, 100)
   t.log(`25 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
+  time = await execute(childs, 25, 100, {useProcFiles: true})
+  t.log(`(procfile) 25 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
   time = await execute(childs, 50, 100)
   t.log(`50 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
+  time = await execute(childs, 50, 100, {useProcFiles: true})
+  t.log(`(procfile) 50 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
   time = await execute(childs, 100, 100)
   t.log(`100 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
+
+  time = await execute(childs, 100, 100, {useProcFiles: true})
+  t.log(`(procfile) 100 pid 100 times done in ${time.toFixed(3)} ms (${(1000 * 100 / time).toFixed(3)} op/s)`)
 
   await destroy(childs)