@simonepri I think that I did the bench thing correctly, looks like using procfiles is really faster:  

```

  ✔ bench › should execute the benchmark (12.9s)
    ℹ 1 pid 100 times done in 1325.301 ms (75.455 op/s)
    ℹ (procfile) 1 pid 100 times done in 28.341 ms (3528.501 op/s)
    ℹ 2 pid 100 times done in 1299.832 ms (76.933 op/s)
    ℹ (procfile) 1 pid 100 times done in 52.198 ms (1915.772 op/s)
    ℹ 5 pid 100 times done in 1316.378 ms (75.966 op/s)
    ℹ (procfile) 5 pid 100 times done in 45.734 ms (2186.534 op/s)
    ℹ 10 pid 100 times done in 1354.110 ms (73.849 op/s)
    ℹ (procfile) 10 pid 100 times done in 50.611 ms (1975.843 op/s)
    ℹ 25 pid 100 times done in 1358.462 ms (73.613 op/s)
    ℹ (procfile) 25 pid 100 times done in 99.147 ms (1008.599 op/s)
    ℹ 50 pid 100 times done in 1500.427 ms (66.648 op/s)
    ℹ (procfile) 50 pid 100 times done in 144.820 ms (690.511 op/s)
    ℹ 100 pid 100 times done in 1513.935 ms (66.053 op/s)
    ℹ (procfile) 100 pid 100 times done in 245.986 ms (406.526 op/s)

``` 

May you check this? 

@vmarchaud may fix #64 let me know if you can try the branch.