@simonepri 2018-06-13
If we can remove the usage of ps and it also faster I'm really welcome with that.
But I think we need to take our time to do it correctly.

Actually I was thinking about implementing the readings in c++ using procfiles and then just expose them to node. (A lot less overhead)
But let's first define a coherent way to obtain the readings on all the platforms.