@vmarchaud 2018-06-13
@simonepri We can't use a module with native compilation with pm2 since its heavily used on a lot of platform