@simonepri 2018-06-13
@vmarchaud Yes the first thing is to implement them in js.
Then if the option is feasible, we could think a way to make a native cross-compatible implementation and ship pre-built binaries. (But is just an idea, it would require a lot of work)