@soyuka 2018-06-13
> Anyway, it would just be a c file with some stdio operations (open/close file) nothing involving strange libraries or something like that.

:+1: thought of this as well, it could be an optional dependency to this project so that it would not interfere.