@soyuka 2018-06-13
> If we can remove the usage of ps and it also faster It's cool.
> But let's first define a coherent way to obtain the readings on all the platforms before diving in the implementation.

Just like it was before:
- os x => uses ps
- win => wmic
- rest => procfiles

I'm :+1: to keep an option that gives the ability to switch to using `ps` (if `procfiles` is the default) or to use `procfiles` (is `ps` is the default).