@vmarchaud 2018-06-13
I'm good to go forward with this options, so do we need to modify pm2 to use procfiles by default ?