@simonepri 2018-06-13
> I'm 👍 to keep an option that gives the ability to switch to using ps (if procfiles is the default) or to use procfiles (is ps is the default).

Mhh, IMO making it configurable just adds complexity for the final user.
If we are facing difficulties trying to understand which solution is "correct way" (although, there isn't a really correct way) I cannot imagine an end user that uses this package figuring out which of the two is better. 🙈

The important factors that I can see here are:
- Speed (How many readings we can take in a second?)
- Meaning (What is the meaning of the readings taken?)
- Portability (This method works well on a lot of platforms?)
- Coherence (We do preserve the meaning of the reading across platform and/or various solution on the same platform?)
- Ease of use (How much knowledge is required to use the API?)