@soyuka 2018-06-13
> Speed (How many readings we can take in a second?)

Looks like procfiles are much much faster since I'm keeping fd's open and not using `readFile`

> Meaning (What is the meaning of the readings taken?)

As shown, when it comes to monitoring, procfiles values are much more promising if you want to know "how much cpu was used in the last X seconds". At least this is what @vmarchaud observed.

> Portability (This method works well on a lot of platforms?)

Because os x has no procfiles, we'll keep using `ps` until we can find a way to add some sycalls.

> Coherence (We do preserve the meaning of the reading across platform and/or various solution on the same platform?)

According to my recent tests, readings are inconsistent between ps and procfile if we use our own history. Indeed values will be closer to what `top` does (cf stackoverflow post on the subject).
 
> Ease of use (How much knowledge is required to use the API?)

We need to document that ps `pcpu` is percent cpu usage since process started. We can also document that `procfiles` calls are "usage since last call".