@simonepri 2018-06-13
> We need to document that ps pcpu is percent cpu usage since process started. We can also document that procfiles calls are "usage since last call".
We can obtain the same meaning using ps on macOS using etime and stime.
We have to try to guarantee the same meaning across platform if possible.