@simonepri 2018-06-13
> it's useless for us to get a "realtime" metric that is already aggregated.
But the CPU percentage is an aggregated info by definition.
If you want a real non-aggregated reading for keymetrics you should use`etime` and `stime` values and then compute the percentage on your side.