@vmarchaud 2018-06-13
> But the CPU percentage is an aggregated info by definition.

right, the CPU percentage can't be sended without being aggregated. But there is a difference between an average of 5 seconds and the whole lifetime of the process.

> If you want a real non-aggregated reading for keymetrics you should useetime and stime values and then compute the percentage on your side.

Actually pidusage mainly exist because we needed it for PM2, i don't want to force anyone to implement it a specific way, but if we need to implement it on our side, we will just fork pidusage to works the way we want.