@simonepri 2018-06-13
@soyuka what you think about the change of reporting both `etime` and `stime` inside the output object instead of an aggregated `ctime` value?