@simonepri 2018-06-13
> Actually pidusage mainly exist because we needed it for PM2, i don't want to force anyone to implement it a specific way, but if we need to implement it on our side, we will just fork pidusage to works the way we want.

Yeah, but if possible is better to keep a single and well working module.

> right, the CPU percentage can't be sended without being aggregated. But there is a difference between an average of 5 seconds and the whole lifetime of the process.

Yeah, actually the current behaviour is that it depends on how frequently you perform the reading.
If you perform the reading 1 time/second then you'll obtain the usage in the last second.
The maxage of a reading in the past is 60 seconds. So if you perform a measure 1 time each 5 minutes actually the reading is took twice and you get the cpu usage in the last couple of ms. (The maxage is configurable)