@soyuka 2018-06-13
> @soyuka what you think about the chance of reporting both etime and stime inside the output object instead of an aggregated ctime value?

It's hard right now because we tried to make `ps` work the same on linux and osx (remember the different time formats and the non-presence of some options). If we say that `ps` now works only for `osx` me may improve the `ps` part a lot!

I'll think about that though, I need to read back all our discussions in the last big changes you did!