@simonepri 2018-06-13
That's how I'm imagine we could give the output in the future:
(We may also try to make the output configurable E.G. "I just want info about the cpu")
```js
{
  proc: {
    ppid: 312,         // PPID
    pid: 727,          // PID
    uptime: 6650000,   // ms since the start of the process,
    thcount: 2         // number of threads used by the process
  }
  cpu: {
    pcpu: 15.0,        // percentage since last call (from 0 to 100*vcore)
    utime: 867000,     // user time in ms
    stime: 867000,     // kernel/system time in ms
  },
  mem: {
    pmem: 34.3,        // percentage (from 0 to 100)
    rss: 357306368,    // resident set size in bytes
    vsz: 957306368,    // virtual memory size in bytes
  },
  timestamp: 864000000  // ms since epoch in witch the reading has been taken
}
```

@soyuka @vmarchaud any suggestion about that?