@soyuka 2018-06-13
> That's how I imagine we could give the output in the future

As long as it's backward compatible I see no harm, we're already computing these values so why not!

Should we keep the pcpu since start time?