@simonepri 2018-06-13
> Should we keep the pcpu since start time?

I think that the user expects the other way around.
If they need the percentage since the start they can compute it by just doing:

`etime/(stime + utime)`