@soyuka 2018-06-14
It feels wrong to add the `again` test to ps, it's also slow enough I'm really not sure I want to slow it more :smile: