@simonepri 2018-06-14
It's just the first time
The reading should be correct, no matter if it takes a bit longer.