@soyuka 2018-06-14
It's a bit more complicated then that because we allow multiple pids so we need to recompute all of them :|.